requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
langchain==0.3.27
langchain-groq==0.3.7
chromadb==1.0.16
sentence-transformers==5.1.0
pypdf==5.9.0
langchain_community==0.3.27
ftfy==6.3.1
cleantext==1.1.4
pdfplumber==0.11.7
colorlog==6.9.0
